+++ packages/bottom-sheet/src/components/header/Component.tsx
@@ -2,11 +2,14 @@ import React, { FC, RefObject, useContext, useEffect } from 'react';
 import cn from 'classnames';
 
 import { BaseModalContext } from '@alfalab/core-components-base-modal';
-import { NavigationBar, NavigationBarProps } from '@alfalab/core-components-navigation-bar';
+import {
+    NavigationBarPrivate,
+    NavigationBarPrivateProps,
+} from '@alfalab/core-components-navigation-bar-private';
 
 import styles from './index.module.css';
 
-export type HeaderProps = Omit<NavigationBarProps, 'view' | 'size'> & {
+export type HeaderProps = Omit<NavigationBarPrivateProps, 'view' | 'size'> & {
     headerRef: RefObject<HTMLDivElement>;
     headerOffset: number;
 };
@@ -34,7 +37,7 @@ export const Header: FC<HeaderProps> = ({
     const hasContent = Boolean(title || children);
 
     return (
-        <NavigationBar
+        <NavigationBarPrivate
             {...restProps}
             ref={headerRef}
             title={title}
@@ -49,6 +52,6 @@ export const Header: FC<HeaderProps> = ({
             contentClassName={cn(styles.title)}
         >
             {children}
-        </NavigationBar>
+        </NavigationBarPrivate>
     );
 };