+++ packages/input/src/components/clear-button/Component.tsx
@@ -17,11 +17,17 @@ interface ClearButtonProps {
     disabled?: boolean;
     onClick: (event: React.MouseEvent<HTMLButtonElement>) => void;
     colors: 'default' | 'inverted';
+    dataTestId?: string;
 }
 
 const preventDefault = (e: React.UIEvent<HTMLElement>) => e.preventDefault();
 
-export const ClearButton: React.FC<ClearButtonProps> = ({ colors, disabled, onClick }) => (
+export const ClearButton: React.FC<ClearButtonProps> = ({
+    colors,
+    disabled,
+    onClick,
+    dataTestId,
+}) => (
     <Button
         type='button'
         view='text'
@@ -31,6 +37,7 @@ export const ClearButton: React.FC<ClearButtonProps> = ({ colors, disabled, onCl
         onClick={onClick}
         tabIndex={-1}
         onMouseDown={preventDefault}
+        dataTestId={dataTestId}
     >
         <CrossCircleMIcon className={cn(styles.clearIcon, colorStyles[colors].clearIcon)} />
     </Button>