+++ packages/navigation-bar/src/shared/index.ts
@@ -1,2 +1,3 @@
-export * from '../components/closer';
-export * from '../components/back-arrow-addon';
+export * from '../components/action-icon-addon';
+export * from '../components/action-text-addon';
+export { getNavigationBarTestIds } from '../utils';