+++ packages/product-cover/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './overlay';
+export * from './text';
+export * from './image-card';