+++ packages/toast-plate/shared/index.ts
@@ -1,0 +1,1 @@
+export { getBaseToastPlateTestIds } from '../src/components/base-toast-plate/utils/getBaseToastPlateTestIds';