+++ packages/toast-plate/src/components/base-toast-plate/types/unsafeBadgeProps.ts
@@ -1,0 +1,6 @@
+/**
+ * @deprecated Данный тип не будет поддерживаться в будущих версиях
+ * Используйте StatusBadgeProps['view']
+ */
+// eslint-disable-next-line @typescript-eslint/naming-convention
+export type unsafe_BadgeProps = 'negative' | 'positive' | 'attention';