+++ packages/action-button/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.6.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.10.0
+
 ## 1.6.1
 
 ### Patch Changes