+++ packages/alert/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 4.1.22
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   plate@7.3.8
+
+## 4.1.21
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   plate@7.3.7
+
 ## 4.1.20
 
 ### Patch Changes