+++ packages/calendar/CHANGELOG.md
@@ -1,5 +1,28 @@
 # Change Log
 
+## 7.6.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.10.0
+    -   button@11.4.3
+    -   modal@9.3.1
+    -   icon-button@6.5.7
+
+## 7.6.2
+
+### Patch Changes
+
+<sup><time>24.05.2024</time></sup>
+
+### [#1193](https://github.com/core-ds/core-components/pull/1193)
+
+-   Добавили dataTestId к PeriodSlider
+
+-   Обновлены зависимости
+    -   modal@9.3.0
+
 ## 7.6.1
 
 ### Patch Changes