+++ packages/password-input/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 5.2.7
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@14.4.3
+    -   icon-button@6.5.7
+
+## 5.2.6
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@14.4.2
+
 ## 5.2.5
 
 ### Patch Changes