+++ packages/tooltip/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 7.3.12
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   bottom-sheet@6.15.1
+    -   button@11.4.3
+
+## 7.3.11
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   bottom-sheet@6.15.0
+
 ## 7.3.10
 
 ### Patch Changes